# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API CashWarden",
        "description": "Toutes les requêtes API CashWarden",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/limitedBudgets": {
            "get": {
                "tags": [
                    "LimitedBudgets"
                ],
                "summary": "Permet d'obtenir la liste des budgets limités",
                "description": "Retourne la liste des budgets limités",
                "operationId": "getLimitedBudgets",
                "responses": {
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "limitedBudgets": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "amount": {
                                                        "type": "string",
                                                        "example": "5800"
                                                    },
                                                    "previsionDate": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2022-03-01"
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2022-03-11T21:44:23.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2022-03-14T23:30:22.000000Z"
                                                    },
                                                    "isLatestLimitedBudgetOfUser": {
                                                        "type": "boolean",
                                                        "example": "true"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "LimitedBudgets"
                ],
                "summary": "Permet de créer un budget limité",
                "description": "Crée un budget limité",
                "operationId": "postLimitedBudgets",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreLimitedBudgetRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "limitedBudgets": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "amount": {
                                                        "type": "string",
                                                        "example": "2000.0"
                                                    },
                                                    "previsionDate": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2022-03-01"
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2022-03-11T21:44:23.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2022-03-14T23:30:22.000000Z"
                                                    },
                                                    "isLatestLimitedBudgetOfUser": {
                                                        "type": "boolean",
                                                        "example": "true"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/limitedBudgets/{id}": {
            "put": {
                "tags": [
                    "LimitedBudgets"
                ],
                "summary": "Permet de modifier un budget limité",
                "description": "Modifie un budget limité",
                "operationId": "putLimitedBudgets",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "limited budget id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateLimitedBudgetRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "amount": {
                                            "type": "string",
                                            "example": "5800"
                                        },
                                        "previsionDate": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2022-03-01"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2022-03-11T21:44:23.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2022-03-14T23:30:22.000000Z"
                                        },
                                        "isLatestLimitedBudgetOfUser": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreLimitedBudgetRequest": {
                "properties": {
                    "amount": {
                        "type": "number",
                        "example": "2000.0"
                    },
                    "previsionDate": {
                        "type": "string",
                        "format": "date",
                        "example": "2022-03-09"
                    }
                },
                "type": "object"
            },
...
            "UpdateLimitedBudgetRequest": {
                "properties": {
                    "amount": {
                        "type": "number",
                        "example": "2000.0"
                    },
                    "previsionDate": {
                        "type": "string",
                        "format": "date",
                        "example": "2022-03-09"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}